QAbstractEventDispatcher: change event dispatcher timer interval to qint64
Resolve remaining Qt6 TODO

[ChangeLog][QtCore][QAbstractEventDispatcher] The signature of the abstract virtual registerTime function now takes a qint64 value for the interval parameter.

+++ src/corelib/kernel/qabstracteventdispatcher.cpp
@@ -241,7 +241,7 @@ QAbstractEventDispatcher *QAbstractEventDispatcher::instance(QThread *thread)
     Registers a timer with the specified \a interval and \a timerType for the
     given \a object and returns the timer id.
 */
-int QAbstractEventDispatcher::registerTimer(int interval, Qt::TimerType timerType, QObject *object)
+int QAbstractEventDispatcher::registerTimer(qint64 interval, Qt::TimerType timerType, QObject *object)
 {
     int id = QAbstractEventDispatcherPrivate::allocateTimerId();
     registerTimer(id, interval, timerType, object);
