+++ src/corelib/kernel/qeventdispatcher_glib_p.h
@@ -75,7 +75,7 @@ public:
     void registerSocketNotifier(QSocketNotifier *socketNotifier) final;
     void unregisterSocketNotifier(QSocketNotifier *socketNotifier) final;
 
-    void registerTimer(int timerId, int interval, Qt::TimerType timerType, QObject *object) final;
+    void registerTimer(int timerId, qint64 interval, Qt::TimerType timerType, QObject *object) final;
     bool unregisterTimer(int timerId) final;
     bool unregisterTimers(QObject *object) final;
     QList<TimerInfo> registeredTimers(QObject *object) const final;
